# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Roommate Chore & Expense Roulette API",
    "version": "1.0.0",
    "description": "# 🎲 Roommate Chore & Expense Roulette API\n\nProduction-grade FastAPI backend featuring specialized Data Structures & Algorithms (DSA):\n- **Chore Rotation**: Circular Queue (Ring Buffer) with O(1) modulo turn shift and sarcastic alerts\n- **Expense Ledger**: Hash Tables with O(1) net balance computations and strict zero-sum conservation\n- **Debt Simplification**: Min-Cash-Flow Greedy Directed Graph Algorithm reducing O(N^2) debts to <= N - 1 vectors\n- **Activity Undo Stack**: LIFO Stack for idempotent database state reversal\n- **Identity & Keys**: RFC 9562 UUIDv7 monotonic primary keys and Ed25519 asymmetric cryptography",
    "contact": {
      "name": "Roulette Engineering Team",
      "email": "engineering@roommate-roulette.local",
      "url": "https://github.com/roommate-roulette"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local Development Server"
    },
    {
      "url": "https://api.roommate-roulette.local",
      "description": "Production Gateway"
    }
  ],
...
  "paths": {
...
    "/api/v1/chores/{household_id}": {
      "get": {
        "tags": ["Chores & Circular Queue"],
        "summary": "List household chores & active assignees",
        "description": "Returns all chores for the household along with current circular queue pointer assignees.",
        "parameters": [
          {
            "name": "household_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of household chores"
          }
        }
      }
    },
    "/api/v1/chores/": {
      "post": {
        "tags": ["Chores & Circular Queue"],
        "summary": "Create chore assigned to circular queue head",
        "description": "Registers a new chore duty and initializes its pointer at the head of the circular queue.",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "201": {
            "description": "Chore successfully created"
          }
        }
      }
    },
    "/api/v1/chores/{chore_id}/rotate": {
      "post": {
        "tags": ["Chores & Circular Queue"],
        "summary": "Rotate circular queue pointer to next roommate O(1)",
        "description": "Executes O(1) turn shift via `(current_index + 1) % N`. Triggers dynamic Sarcastic Alert if user attempts to skip.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "chore_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Queue rotated successfully"
          },
          "400": {
            "description": "Premature skip with Sarcastic Alert"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "PyJWT Bearer token"
      },
      "Ed25519Signature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Ed25519-Signature",
        "description": "64-byte Ed25519 signature over server challenge nonce"
      },
      "Ed25519PublicKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Ed25519-Public-Key",
        "description": "32-byte Ed25519 hex public key"
      }
    },
...
  }
}